feat: Introduce StringIntervalUtility for managing string intervals
- Added StringIntervalUtility class to handle string intervals based on specified keys.
- Implemented methods for transforming and merging intervals.
- Created utility methods for finding and flattening string intervals.

refactor: Organize utility exports in act_flutter_utility

- Updated exports in act_flutter_utility.dart to point to new utility files.
- Removed deprecated text_utility.dart and organized utility functions into separate files.

feat: Implement ColorsUtility for color manipulation

- Added ColorsUtility class with methods for lighting and darkening colors.
- Implemented gradient color interpolation method.

feat: Create OverlayUtility for managing overlays

- Introduced OverlayUtility class to display widgets above the current view.
- Implemented overlay management with animation support.

feat: Develop TextUtility for text manipulation

- Added TextUtility class for highlighting and replacing text with widgets.
- Integrated StringIntervalUtility for interval management in text highlighting.

feat: Add UrlLauncherUtility for URL handling

- Created UrlLauncherUtility class to open URLs in the app browser.

feat: Introduce WidgetUtility for widget-related functions

- Added WidgetUtility class with methods for size calculations and widget management.
- Implemented utility methods for handling icons and conditional scroll views.
