+++ act_dart_utility/lib/act_dart_utility.dart
@@ -25,6 +25,7 @@ export 'src/models/result_with_bool_status.dart';
 export 'src/models/result_with_required_value.dart';
 export 'src/models/result_with_status.dart';
 export 'src/models/status_with_extra_info.dart';
+export 'src/models/string_interval.dart';
 export 'src/models/updated_model_event.dart';
 export 'src/models/updated_unique_model_event.dart';
 export 'src/models/value_keeper.dart';
@@ -51,6 +52,7 @@ export 'src/utilities/loop_utility.dart';
 export 'src/utilities/map_utility.dart';
 export 'src/utilities/num_utility.dart';
 export 'src/utilities/path_utility.dart';
+export 'src/utilities/string_interval_utility.dart';
 export 'src/utilities/string_list_utility.dart';
 export 'src/utilities/string_utility.dart';
 export 'src/utilities/type_utility.dart';
