+++ act_flutter_utility/lib/act_flutter_utility.dart
@@ -10,7 +10,7 @@ export 'src/blocs/bloc_event_for_mixin.dart';
 export 'src/blocs/bloc_for_mixin.dart';
 export 'src/blocs/bloc_state_for_mixin.dart';
 export 'src/blocs/mixin_generic_loading_state.dart';
-export 'src/colors_util.dart';
+export 'src/utilities/colors_utility.dart';
 export 'src/graphical_assets/graphical_asset.dart';
 export 'src/graphical_assets/icon_asset.dart';
 export 'src/graphical_assets/png_asset.dart';
@@ -19,11 +19,11 @@ export 'src/models/act_tab_bar_config.dart';
 export 'src/models/banner_information_model.dart';
 export 'src/models/extended_text_style.dart';
 export 'src/models/text_span_config.dart';
-export 'src/overlay_util.dart';
-export 'src/text_utility.dart';
+export 'src/utilities/overlay_utility.dart';
+export 'src/utilities/text_utility.dart';
 export 'src/types/banner_information_type.dart';
 export 'src/types/single_child_scroll_view_type.dart';
-export 'src/url_launcher_utility.dart';
+export 'src/utilities/url_launcher_utility.dart';
 export 'src/widget/act_visibility_detector.dart';
 export 'src/widget/action_error_retry/action_error_retry_theme.dart';
 export 'src/widget/action_error_retry/action_error_retry_widget.dart';
@@ -36,4 +36,4 @@ export 'src/widget/scroll_view/single_expandable_child_scroll_view.dart';
 export 'src/widget/snack_bar/act_snack_bar.dart';
 export 'src/widget/tab_bars/abs_simple_tab_bar.dart';
 export 'src/widget/tab_bars/mixin_simple_tab_bar_state.dart';
-export 'src/widget_utils.dart';
+export 'src/utilities/widget_utility.dart';
