+++ act_flutter_utility/lib/src/utilities/url_launcher_utility.dart
